Rename envVar package to overrides
An upcoming change will allow OLM users to set the Volume and
VolumeMounts in operator  deployments via the subscription config field.
There is a fair amount of code that resides in the envvar package that
could be reused.
