+++ pkg/controller/operators/olm/overrides/config.go
@@ -1,4 +1,4 @@
-package envvar
+package overrides
 
 import (
 	"fmt"
