+++ pkg/controller/operators/olm/overrides/initializer.go
@@ -1,4 +1,4 @@
-package envvar
+package overrides
 
 import (
 	"fmt"
