+++ pkg/controller/operators/olm/overrides/inject.go
@@ -1,4 +1,4 @@
-package envvar
+package overrides
 
 import (
 	"errors"
