+++ pkg/package-server/provider/manifests/prometheus/0.14.0/alertmanager.crd.yaml
@@ -1436,11 +1436,11 @@ spec:
                     pending:
                       description: Pending is a list of initializers that must execute
                         in order before this object is visible. When the last pending
-                        initializer is removed, and no failing result is set, the
+                        overrides is removed, and no failing result is set, the
                         initializers struct will be set to nil and the object is considered
                         as initialized and visible to all clients.
                       items:
-                        description: Initializer is information about an initializer
+                        description: Initializer is information about an overrides
                           that has not yet completed.
                         properties:
                           name:
@@ -1937,12 +1937,12 @@ spec:
                             pending:
                               description: Pending is a list of initializers that
                                 must execute in order before this object is visible.
-                                When the last pending initializer is removed, and
+                                When the last pending overrides is removed, and
                                 no failing result is set, the initializers struct
                                 will be set to nil and the object is considered as
                                 initialized and visible to all clients.
                               items:
-                                description: Initializer is information about an initializer
+                                description: Initializer is information about an overrides
                                   that has not yet completed.
                                 properties:
                                   name:
