+++ pkg/package-server/provider/manifests/prometheus/0.15.0/prometheus.crd.yaml
@@ -1524,11 +1524,11 @@ spec:
                     pending:
                       description: Pending is a list of initializers that must execute
                         in order before this object is visible. When the last pending
-                        initializer is removed, and no failing result is set, the
+                        overrides is removed, and no failing result is set, the
                         initializers struct will be set to nil and the object is considered
                         as initialized and visible to all clients.
                       items:
-                        description: Initializer is information about an initializer
+                        description: Initializer is information about an overrides
                           that has not yet completed.
                         properties:
                           name:
@@ -2439,12 +2439,12 @@ spec:
                             pending:
                               description: Pending is a list of initializers that
                                 must execute in order before this object is visible.
-                                When the last pending initializer is removed, and
+                                When the last pending overrides is removed, and
                                 no failing result is set, the initializers struct
                                 will be set to nil and the object is considered as
                                 initialized and visible to all clients.
                               items:
-                                description: Initializer is information about an initializer
+                                description: Initializer is information about an overrides
                                   that has not yet completed.
                                 properties:
                                   name:
